refactor(core): lift assigner pipeline to a leaf module, breaking the last import cycle
Move the v-register assigner pipeline (assignKey, the default/consumer
assigner tags + predicates, the transforms runner, kickoffAsyncTransform,
applyCoerce, getModelAssigner, fireAssigner, setAssignFunction) out of
directive.ts into a new leaf module core/assigner-pipeline.ts (Stage C/8
of the fallow healthy-refactor roadmap; the assigner-pipeline cycle-break
signed off before coding).

This removes the residual directive.ts <-> directive-file.ts import cycle
(the last one): directive-file.ts now imports fireAssigner /
setAssignFunction from the leaf rather than from directive.ts. The
surviving edge directive -> directive-file (vRegisterFile) is
one-directional. assigner-pipeline imports only leaf modules
(register-protocol, types, vue-shared-shim, dev, paths, vue) — no
directive / directive-* back-edge.

Mechanical, behavior-preserving: bodies moved byte-for-byte (seeded via
file copy, never retyped). assignKey is re-exported from directive.ts so
the public `attaform` entry (src/index.ts) is byte-identical, and it
stays Symbol.for('attaform:assign-key') so consumer el[assignKey] round-
trips. The directive definitions, listeners, shouldBailListener,
applyElementCoerce, writeLastTypedForm, and composition handlers stay in
directive.ts.

Cycle count -> 0 (structurally: no import path returns to directive.ts;
make fallow 84 B -> 86 A, no circular-dependency findings). 392 register
/ directive / transform / coerce / file tests pass unchanged; tsc clean.
